+++ src/pipeline.rs
@@ -3,7 +3,7 @@ use std::{fs, path::PathBuf};
 use itertools::Itertools;
 
 use crate::{
-    alignment::{Alignment, AlignmentData, Strand, VecMap},
+    alignment::{Alignment, AlignmentData, Strand},
     alphabet::UTF8_TO_DIGITAL_NUCLEOTIDE,
     chunks::ProximityGroup,
     confidence::confidence,
@@ -46,284 +46,6 @@ pub fn run_assembly_pipeline(
     confidence(&mut confidence_matrix);
 
     let confidence_avg_by_row = windowed_confidence_slow(&mut confidence_matrix);
-
-    let mut query_ids: Vec<usize> = group
-        .alignments
-        .iter()
-        .map(|a| a.query_id)
-        .unique()
-        .collect();
-
-    query_ids.sort();
-
-    struct AlignmentTuple<'a> {
-        row_idx: usize,
-        alignment: &'a Alignment,
-        confidence: f64,
-    }
-
-    #[derive(PartialEq, Clone, Debug)]
-    struct Edge {
-        node_idx: usize,
-        weight: f64,
-    }
-
-    query_ids
-        .iter()
-        // grab the alignments for this ID
-        .map(|id| {
-            (
-                id,
-                group
-                    .alignments
-                    .iter()
-                    // enumerate to get the row
-                    .enumerate()
-                    .map(|(idx, ali)| AlignmentTuple {
-                        row_idx: idx + 1,
-                        alignment: ali,
-                        confidence: confidence_avg_by_row[idx + 1],
-                    })
-                    .filter(|t| t.alignment.query_id == *id),
-            )
-        })
-        .for_each(|(query_id, alignments)| {
-            // split the forward and reverse stranded alignments
-            let (fwd, rev): (Vec<_>, Vec<_>) = alignments
-                .into_iter()
-                .partition(|a| a.alignment.strand == Strand::Forward);
-
-            if fwd.is_empty() {
-                return;
-            }
-
-            fwd.iter().zip(fwd.iter().skip(1)).for_each(|(a, b)| {
-                debug_assert!(a.alignment.target_start <= b.alignment.target_start);
-            });
-
-            let mut graph: Vec<Vec<Edge>> = vec![vec![]; fwd.len()];
-            println!();
-            (0..fwd.len()).for_each(|a_idx| {
-                (a_idx + 1..fwd.len()).for_each(|b_idx| {
-                    let a = &fwd[a_idx];
-                    let b = &fwd[b_idx];
-
-                    if a.alignment == b.alignment {
-                        return;
-                    }
-
-                    let within_target_distance = b
-                        .alignment
-                        .target_start
-                        .saturating_sub(a.alignment.target_end)
-                        < args.target_join_distance;
-
-                    let consensus_distance =
-                        b.alignment.query_start as isize - a.alignment.query_end as isize;
-
-                    // TODO: PARAMETERIZE THIS
-                    let consensus_colinear = consensus_distance > -20;
-
-                    if within_target_distance && consensus_colinear {
-                        graph[a_idx].push(Edge {
-                            node_idx: b_idx,
-                            weight: consensus_distance.abs() as f64,
-                        });
-                        graph[b_idx].push(Edge {
-                            node_idx: a_idx,
-                            weight: consensus_distance.abs() as f64,
-                        });
-                    }
-                });
-            });
-
-            graph
-                .iter()
-                .enumerate()
-                .for_each(|(i, v)| println!("{i} ({:.4}): {v:?}", fwd[i].confidence));
-            // take all indices of the remaining alignment tuples
-            let mut remaining_indices = (0..fwd.len()).collect_vec();
-
-            // sort them by their respective confidence values
-            remaining_indices.sort_by(|a, b| {
-                fwd[*a]
-                    .confidence
-                    .partial_cmp(&fwd[*b].confidence)
-                    .expect("failed to compare confidence values")
-            });
-
-            // we'll keep track of the assembled
-            // alignments as vectors of those indices
-            let mut assemblies: VecMap<Vec<usize>> = VecMap::new();
-
-            let mut assembly: Vec<usize>;
-
-            // while we have alignments left to place
-            while !remaining_indices.is_empty() {
-                // start an assembly with the highest remaining confidence
-                assembly = vec![remaining_indices.pop().unwrap()];
-                loop {
-                    // the alignment we started with is the
-                    // one that's going to maintain edges
-                    let root_idx = assembly[0];
-                    let root_edges = &graph[root_idx];
-
-                    // find the edge with the minimum weight
-                    let join_edge = match root_edges.iter().min_by(|a, b| {
-                        a.weight
-                            .partial_cmp(&b.weight)
-                            .expect("failed to compare edge weights")
-                    }) {
-                        Some(edge) => edge,
-                        // if we don't have an edge,
-                        // we're done with this assembly
-                        None => break,
-                    };
-
-                    let join_idx = join_edge.node_idx;
-
-                    // add this alignment to the assembly
-                    assembly.push(join_idx);
-                    // and remove it from further consideration
-
-                    // println!("{root_idx} {join_idx}, {remaining_indices:?}");
-                    // println!("{edges:?}");
-                    // println!();
-
-                    match remaining_indices.iter().position(|&idx| idx == join_idx) {
-                        Some(pos) => remaining_indices.remove(pos),
-                        None => panic!("failed to remove index from remaining"),
-                    };
-
-                    // now take it's edges
-                    let join_edges = &graph[join_idx];
-
-                    let root_edge_indices = root_edges.iter().map(|e| e.node_idx).collect_vec();
-
-                    let compatible_indices = join_edges
-                        .iter()
-                        .filter(|e| root_edge_indices.contains(&e.node_idx))
-                        .map(|e| e.node_idx)
-                        .collect_vec();
-
-                    let root_compatible_edges = root_edges
-                        .iter()
-                        .filter(|e| compatible_indices.contains(&e.node_idx))
-                        .collect_vec();
-
-                    let join_compatible_edges = join_edges
-                        .iter()
-                        .filter(|e| compatible_indices.contains(&e.node_idx))
-                        .collect_vec();
-
-                    debug_assert_eq!(root_compatible_edges.len(), join_compatible_edges.len());
-
-                    let new_root_edges = root_compatible_edges
-                        .iter()
-                        .zip(join_compatible_edges)
-                        .map(|(r, j)| Edge {
-                            node_idx: r.node_idx,
-                            weight: r.weight.min(j.weight),
-                        })
-                        .collect_vec();
-
-                    graph[root_idx] = new_root_edges;
-
-                    if !compatible_indices.is_empty() {
-                        println!("{}", compatible_indices.len());
-                    }
-
-                    graph[join_idx] = vec![];
-                }
-            }
-
-            // ---------
-            // viz stuff
-            // ---------
-
-            let fwd_consensus_ali = fwd
-                .iter()
-                .map(|t| {
-                    serde_json::json!({
-                        "id": t.row_idx.to_string(),
-                        "start": t.alignment.query_start,
-                        "end": t.alignment.query_end,
-                        "strand": t.alignment.strand,
-                        "conf": t.confidence,
-                    })
-                })
-                .collect::<serde_json::Value>();
-
-            let fwd_target_ali = fwd
-                .iter()
-                .map(|t| {
-                    serde_json::json!({
-                        "id": t.row_idx.to_string(),
-                        "start": t.alignment.target_start,
-                        "end": t.alignment.target_end,
-                        "strand": t.alignment.strand,
-                        "conf": t.confidence,
-                    })
-                })
-                .collect::<serde_json::Value>();
-
-            let links = fwd
-                .iter()
-                .enumerate()
-                .flat_map(|(a_idx, a_t)| {
-                    graph[a_idx]
-                        .iter()
-                        .map(|edge| &fwd[edge.node_idx])
-                        .map(|b_t| {
-                            let start = (a_t.alignment.target_start + a_t.alignment.target_end) / 2;
-                            let end = (b_t.alignment.target_start + b_t.alignment.target_end) / 2;
-                            serde_json::json!({
-                                "id": format!("{}-{}", a_t.row_idx, b_t.row_idx),
-                                "start": start,
-                                "end": end,
-                            })
-                        })
-                })
-                .collect::<Vec<_>>();
-
-            let query_id_idx = query_ids
-                .iter()
-                .position(|id| id == query_id)
-                .expect("failed to find query_id");
-
-            let next_idx = query_id_idx + 1;
-
-            let next = if next_idx < query_ids.len() {
-                query_ids[next_idx]
-            } else {
-                query_ids[0]
-            };
-
-            let prev_idx = query_id_idx as isize - 1;
-            let prev = if prev_idx > 0 {
-                query_ids[prev_idx as usize]
-            } else {
-                query_ids[query_ids.len() - 1]
-            };
-
-            let fwd_data = serde_json::json!({
-                "consensus": fwd_consensus_ali,
-                "target": fwd_target_ali,
-                "links": links,
-                "prev": prev,
-                "next": next,
-                "suffix": "fwd",
-            });
-
-            let html_path = viz_path.join(format!("{}-fwd.html", query_id));
-
-            write_soda_html(
-                &fwd_data,
-                "./fixtures/soda/assembly-template.html",
-                "./fixtures/soda/assembly.js",
-                html_path,
-            );
-        });
 }
 
 pub fn run_pipeline(
